requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
fdb==2.0.2
pydantic-settings>=2.5.2
fastapi-mcp>=0.2.0
python-dotenv>=1.0.0
pydantic-ai>=0.1.0
openai>=1.12.0